third_party/WebKit/LayoutTests/editing/execCommand/insert-paragraph-into-table.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="sample" contenteditable>
<table><td style="display:inline-block"><ruby><rt>a</rt></ruby></td></table>
</div>
<div id="log"></div>
<script>
test(function() {
    var selection = window.getSelection();
    var sample = document.getElementById('sample');
    selection.collapse(sample, 2);
    selection.extend(document.querySelector('rt').firstChild, 1);
    document.execCommand('InsertParagraph');
    assert_equals(sample.innerHTML.replace(/\n/g, ''), '<table><tbody><tr><td style="display:inline-block"><ruby><rt>a</rt></ruby></td></tr><tr><font size="1"><br></font></tr></tbody></table>');
});
</script>
